requirements.txt
# 四级行业双轨筛选系统依赖
pandas>=1.5.0
numpy>=1.23.0
psycopg2-binary>=2.9.0
PyYAML>=6.0
python-dotenv>=0.21.0
plotly>=5.10.0
kaleido>=0.2.1
openpyxl>=3.0.10
scipy>=1.9.0
